perf+fix: O1/O4/O5/O10 processamento + bugs fps/speed/PDF
Escopo extraido de docs/handoffs/next-agent-handout.md (secao B, PR pequeno
de baixo risco, nao toca esqueleto Fase 2/3, golden-file intocado):

- O1: rectify() so nos frames amostrados do background model (Detect.setup),
  nao nos 1200 frames inteiros pra extrair 3 amostras.
- O4: max incremental (np.maximum out=) no lugar de lista + np.max final,
  evita reter O(duracao/frame_block) frames em memoria.
- O5: remove 2o cv2.threshold (no-op matematica confirmada via np.array_equal).
- O10: pula cv2.warpPerspective quando a matriz e identidade (sem pontos
  reais de perspectiva escolhidos).
- fix: Capture.open() nao trunca mais fps com int() (29.97->29 era ~3% de
  vies em toda velocidade calculada a jusante).
- fix: SpeedPlugin ignora pares de frame nao-contiguos (gap por deteccao
  falha nao e mais tratado como passo de 1 frame).
- fix: rotulo do PDF corrigido de "Quantidade de frames" pra "Quadros com
  posicao reconstruida" (nunca foi o total do video).

pytest -m "not gpu": 311 passed (golden inalterado). ruff/mypy limpos.

+++ src/stages/capture/plugin.py
@@ -64,7 +64,7 @@ class DualVideoFileCapture(Plugin):
             cap.release()
         return width, height
 
-    def open(self) -> tuple[int, Iterator[FramePair]]:
+    def open(self) -> tuple[float, Iterator[FramePair]]:
         """Abre os dois vídeos e devolve `(fps, generator)`. Cada chamada abre do
         zero (reabertura é necessária: o Detect faz uma pré-passada por câmera antes
         da passada pareada — ver Detect.setup). O generator para no vídeo mais curto."""
@@ -76,7 +76,7 @@ class DualVideoFileCapture(Plugin):
             top_cap.release()
             raise CaptureError(f"Falha ao abrir vídeo lateral: {self._side_path}")
 
-        fps = int(top_cap.get(cv2.CAP_PROP_FPS))  # fps do topo, como no legado
+        fps = top_cap.get(cv2.CAP_PROP_FPS)  # fps do topo, como no legado — sem truncar (bug: 29.97->29)
 
         def _generator() -> Iterator[FramePair]:
             index = 0
